# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Research Navigator API - Comprehensive scientific intelligence platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Paper {
  id: string;
  semanticScholarId?: string | null;
  openAlexId?: string | null;
  doi?: string | null;
  title: string;
  abstract?: string | null;
  year?: number | null;
  citationCount: number;
  influentialCitationCount: number;
  venue?: string | null;
  keywords: string[];
  tldr?: string | null;
  url?: string | null;
  source: string;
}

export interface Author {
  id: string;
  name: string;
  affiliations: string[];
  citationCount: number;
  paperCount: number;
  hIndex: number;
}

export type PaperDetail = Paper & {
  authors: Author[];
  references: Paper[];
  citations: Paper[];
};

export interface PaperListResponse {
  papers: Paper[];
  total: number;
  limit: number;
  offset: number;
}

export type CollectionRunStatus =
  (typeof CollectionRunStatus)[keyof typeof CollectionRunStatus];

export const CollectionRunStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface CollectionRun {
  id: string;
  topic: string;
  status: CollectionRunStatus;
  sourcesUsed: string[];
  papersCollected: number;
  papersSkipped: number;
  papersDeduplicated: number;
  errorMessage?: string | null;
  startedAt?: string;
  completedAt?: string | null;
}

export type StartCollectionRequestSourcesItem =
  (typeof StartCollectionRequestSourcesItem)[keyof typeof StartCollectionRequestSourcesItem];

export const StartCollectionRequestSourcesItem = {
  semantic_scholar: "semantic_scholar",
  open_alex: "open_alex",
} as const;

export interface StartCollectionRequest {
  /** @minLength 2 */
  topic: string;
  /**
   * @minimum 10
   * @maximum 1000
   */
  limit?: number;
  yearFrom?: number;
  yearTo?: number;
  sources?: StartCollectionRequestSourcesItem[];
}

export interface CitationGraphNode {
  id: string;
  title: string;
  year?: number | null;
  citationCount: number;
  influentialCitationCount: number;
  depth: number;
  isHub: boolean;
  isBridge: boolean;
  cluster?: string | null;
  authors: string[];
}

export interface CitationGraphEdge {
  source: string;
  target: string;
  isInfluential: boolean;
}

export interface ResearchLineage {
  id: string;
  name: string;
  paperIds: string[];
  description: string;
}

export interface CitationGraph {
  nodes: CitationGraphNode[];
  edges: CitationGraphEdge[];
  seedPaperId: string;
  totalNodes: number;
  totalEdges: number;
  lineages: ResearchLineage[];
}

export interface PaperSummary {
  paperId: string;
  title: string;
  summary: string;
  keyContributions: string[];
  methodology: string;
  impact: string;
}

export type KeywordTrendDataPointsItem = {
  year: number;
  count: number;
  tfidfScore: number;
};

export interface KeywordTrend {
  keyword: string;
  dataPoints: KeywordTrendDataPointsItem[];
  growthRate: number;
  peakYear: number;
}

export interface TopAuthor {
  author: Author;
  paperCountInTopic: number;
  totalCitations: number;
}

export interface PaperCluster {
  id: string;
  label: string;
  keywords: string[];
  paperCount: number;
  growthRate: number;
  description: string;
}

export interface TrendAnalysis {
  runId: string;
  topic: string;
  keywordTrends: KeywordTrend[];
  topAuthors: TopAuthor[];
  clusters: PaperCluster[];
  narrativeSummary: string;
  totalPapersAnalyzed: number;
}

export type ResearchGapEvidenceType =
  (typeof ResearchGapEvidenceType)[keyof typeof ResearchGapEvidenceType];

export const ResearchGapEvidenceType = {
  topic_modeling: "topic_modeling",
  citation_network: "citation_network",
  method_problem_matrix: "method_problem_matrix",
} as const;

export interface ResearchGap {
  id: string;
  title: string;
  description: string;
  evidenceType: ResearchGapEvidenceType;
  supportingPaperIds: string[];
  noveltyScore: number;
  impactScore: number;
  feasibilityScore: number;
}

export interface ResearchProposal {
  id: string;
  gapId: string;
  title: string;
  motivation: string;
  researchQuestions: string[];
  methodology: string;
  expectedContributions: string[];
  challenges: string[];
  noveltyScore: number;
  noveltyExplanation?: string | null;
  supportingPaperIds: string[];
  rawText?: string | null;
}

export type DebateClaimEvidenceStrength =
  (typeof DebateClaimEvidenceStrength)[keyof typeof DebateClaimEvidenceStrength];

export const DebateClaimEvidenceStrength = {
  strong_empirical: "strong_empirical",
  indirect_inference: "indirect_inference",
  theoretical_speculation: "theoretical_speculation",
} as const;

export interface DebateClaim {
  claim: string;
  evidenceStrength: DebateClaimEvidenceStrength;
  sourceIds: string[];
}

export interface DebateTurn {
  id: string;
  sessionId: string;
  round: number;
  role: string;
  content: string;
  claims: DebateClaim[];
}

export type DebateSessionStatus =
  (typeof DebateSessionStatus)[keyof typeof DebateSessionStatus];

export const DebateSessionStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
} as const;

export interface DebateSession {
  id: string;
  collectionRunId: string;
  topic: string;
  controversialQuestion: string;
  subTopics: string[];
  status: DebateSessionStatus;
  finalReport?: string | null;
  consensusPoints: string[];
  disagreementPoints: string[];
  openQuestions: string[];
}

export interface ResearchReport {
  runId: string;
  topic: string;
  overview: string;
  graphInsights: string;
  trendsSummary: string;
  gapsSummary: string;
  controversySummary: string;
  recommendations: string[];
  totalPapers: number;
  generatedAt: string;
}

export type ListPapersParams = {
  topic?: string;
  year?: number;
  limit?: number;
  offset?: number;
  sortBy?: ListPapersSortBy;
};

export type ListPapersSortBy =
  (typeof ListPapersSortBy)[keyof typeof ListPapersSortBy];

export const ListPapersSortBy = {
  citationCount: "citationCount",
  publicationDate: "publicationDate",
  year: "year",
} as const;

export type ListCollectionRuns200 = {
  runs: CollectionRun[];
};

export type GetCitationGraphParams = {
  /**
   * @minimum 1
   * @maximum 3
   */
  depth?: number;
  maxNodes?: number;
};

export type GetResearchGaps200 = {
  gaps: ResearchGap[];
};

export type AnalyzeResearchGaps200 = {
  gaps: ResearchGap[];
  message: string;
};

export type GetProposals200 = {
  proposals: ResearchProposal[];
};

export type GenerateProposals200 = {
  proposals: ResearchProposal[];
  message: string;
};

export type GetDebates200 = {
  sessions: DebateSession[];
};

export type GetDebateTurns200 = {
  turns: DebateTurn[];
};