# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Research Navigator API - Comprehensive scientific intelligence platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collection/runs:
    get:
      operationId: listCollectionRuns
      tags: [collection]
      summary: List all collection runs
      responses:
        "200":
          description: List of collection runs
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/CollectionRun"
                required:
                  - runs

    post:
      operationId: startCollection
      tags: [collection]
      summary: Start a new data collection run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartCollectionRequest"
      responses:
        "201":
          description: Collection started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionRun"
  /collection/runs/{id}:
    get:
      operationId: getCollectionRun
      tags: [collection]
      summary: Get collection run status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionRun"
        "404":
          description: Run not found
...
  /debates/{runId}:
    get:
      operationId: getDebates
      tags: [debates]
      summary: Get debate sessions for a run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Debate sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/DebateSession"
                required:
                  - sessions
  /debates/{runId}/start:
    post:
      operationId: startDebate
      tags: [debates]
      summary: Start a debate session
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Debate started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebateSession"
  /debates/sessions/{sessionId}/turns:
    get:
      operationId: getDebateTurns
      tags: [debates]
      summary: Get debate turns for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Debate turns
          content:
            application/json:
              schema:
                type: object
                properties:
                  turns:
                    type: array
                    items:
                      $ref: "#/components/schemas/DebateTurn"
                required:
                  - turns
...
components:
  schemas:
...
    CollectionRun:
      type: object
      properties:
        id:
          type: string
        topic:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        sourcesUsed:
          type: array
          items:
            type: string
        papersCollected:
          type: integer
        papersSkipped:
          type: integer
        papersDeduplicated:
          type: integer
        errorMessage:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - topic
        - status
        - sourcesUsed
        - papersCollected
        - papersSkipped
        - papersDeduplicated
    StartCollectionRequest:
      type: object
      properties:
        topic:
          type: string
          minLength: 2
        limit:
          type: integer
          default: 200
          minimum: 10
          maximum: 1000
        yearFrom:
          type: integer
        yearTo:
          type: integer
        sources:
          type: array
          items:
            type: string
            enum: [semantic_scholar, open_alex]
          default: [semantic_scholar, open_alex]
      required:
        - topic
...
    DebateClaim:
      type: object
      properties:
        claim:
          type: string
        evidenceStrength:
          type: string
          enum: [strong_empirical, indirect_inference, theoretical_speculation]
        sourceIds:
          type: array
          items:
            type: string
      required:
        - claim
        - evidenceStrength
        - sourceIds
    DebateTurn:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        round:
          type: integer
        role:
          type: string
        content:
          type: string
        claims:
          type: array
          items:
            $ref: "#/components/schemas/DebateClaim"
      required:
        - id
        - sessionId
        - round
        - role
        - content
        - claims
    DebateSession:
      type: object
      properties:
        id:
          type: string
        collectionRunId:
          type: string
        topic:
          type: string
        controversialQuestion:
          type: string
        subTopics:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, running, completed]
        finalReport:
          type: string
          nullable: true
        consensusPoints:
          type: array
          items:
            type: string
        disagreementPoints:
          type: array
          items:
            type: string
        openQuestions:
          type: array
          items:
            type: string
      required:
        - id
        - collectionRunId
        - topic
        - controversialQuestion
        - subTopics
        - status
        - consensusPoints
        - disagreementPoints
        - openQuestions
...