# lib/api-zod/src/generated/types/collectionRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Research Navigator API - Comprehensive scientific intelligence platform
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionRunStatus } from "./collectionRunStatus";

export interface CollectionRun {
  id: string;
  topic: string;
  status: CollectionRunStatus;
  sourcesUsed: string[];
  papersCollected: number;
  papersSkipped: number;
  papersDeduplicated: number;
  errorMessage?: string | null;
  startedAt?: Date;
  completedAt?: Date | null;
}