# lib/api-zod/src/generated/types/debateClaim.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Research Navigator API - Comprehensive scientific intelligence platform
 * OpenAPI spec version: 0.1.0
 */
import type { DebateClaimEvidenceStrength } from "./debateClaimEvidenceStrength";

export interface DebateClaim {
  claim: string;
  evidenceStrength: DebateClaimEvidenceStrength;
  sourceIds: string[];
}